requirements.txt
fastmcp
chromadb
ollama
pymupdf
python-dotenv
